Adds fixes for GUID and makes changes to README
Updates the message_parse.pl file to fix a bug with netcat not listening on the desired port.
Fixes a bug where __message__ file is empty, and perl returns an undefined json error.
message_parse.pl now tracks GUIDs of objects that scripts are attached to, which allows for saving scripts back to Tabletop Simulator.